refactor(android): move classes in subfolders
