+++ packages/firestore/android/src/main/java/io/capawesome/capacitorjs/plugins/firebase/firestore/FirebaseFirestore.java
@@ -8,20 +8,20 @@ import com.google.firebase.firestore.Filter;
 import com.google.firebase.firestore.ListenerRegistration;
 import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.SetOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.AddCollectionSnapshotListenerOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.AddDocumentOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.AddDocumentResult;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.AddDocumentSnapshotListenerOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.DeleteDocumentOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.GetCollectionGroupOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.GetCollectionOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.GetCollectionResult;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.GetDocumentOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.GetDocumentResult;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.QueryCompositeFilterConstraint;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.RemoveSnapshotListenerOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.SetDocumentOptions;
-import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.UpdateDocumentOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.AddCollectionSnapshotListenerOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.AddDocumentOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.results.AddDocumentResult;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.AddDocumentSnapshotListenerOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.DeleteDocumentOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.GetCollectionGroupOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.GetCollectionOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.results.GetCollectionResult;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.GetDocumentOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.results.GetDocumentResult;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.constraints.QueryCompositeFilterConstraint;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.RemoveSnapshotListenerOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.SetDocumentOptions;
+import io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options.UpdateDocumentOptions;
 import io.capawesome.capacitorjs.plugins.firebase.firestore.interfaces.EmptyResultCallback;
 import io.capawesome.capacitorjs.plugins.firebase.firestore.interfaces.NonEmptyResultCallback;
 import io.capawesome.capacitorjs.plugins.firebase.firestore.interfaces.QueryNonFilterConstraint;
