+++ packages/firestore/android/src/main/java/io/capawesome/capacitorjs/plugins/firebase/firestore/classes/constraints/QueryLimitConstraint.java
@@ -1,4 +1,4 @@
-package io.capawesome.capacitorjs.plugins.firebase.firestore.classes;
+package io.capawesome.capacitorjs.plugins.firebase.firestore.classes.constraints;
 
 import com.getcapacitor.JSObject;
 import io.capawesome.capacitorjs.plugins.firebase.firestore.interfaces.QueryNonFilterConstraint;
