+++ packages/firestore/android/src/main/java/io/capawesome/capacitorjs/plugins/firebase/firestore/classes/options/AddDocumentOptions.java
@@ -1,16 +1,16 @@
-package io.capawesome.capacitorjs.plugins.firebase.firestore.classes;
+package io.capawesome.capacitorjs.plugins.firebase.firestore.classes.options;
 
 import com.getcapacitor.JSObject;
 import io.capawesome.capacitorjs.plugins.firebase.firestore.FirebaseFirestoreHelper;
 import java.util.Map;
 import org.json.JSONException;
 
-public class UpdateDocumentOptions {
+public class AddDocumentOptions {
 
     private String reference;
     private Map<String, Object> data;
 
-    public UpdateDocumentOptions(String reference, JSObject data) throws JSONException {
+    public AddDocumentOptions(String reference, JSObject data) throws JSONException {
         this.reference = reference;
         this.data = FirebaseFirestoreHelper.createHashMapFromJSObject(data);
     }
